backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic==0.28.0
httpx==0.27.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
cloudinary==1.40.0
reportlab==4.2.0
qrcode[pil]==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.1
alembic==1.13.1
aiosqlite==0.20.0
